Add keypair_util_from_path helper
